requirements.txt
flask
scikit-learn
numpy
python-docx
gunicorn
google-generativeai
